# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Binder Public API",
    "version": "1.2.0",
    "description": "Read-only access to Binder's own value-added layer: public binders (holdings and valuation), a binder's price history, public trade lists, and Binder's normalized card catalog.\n\nThe catalog covers every card tracked in Binder, in our normalized cross-game model, with a per-card price history built from our own daily snapshots. It is enough to build a collection app of your own on top of. This API deliberately does not expose a live upstream card or price search passthrough. Card data and pricing come from Scryfall (MTG), PokéWallet (Pokémon), and OPTCG API (One Piece); every response includes an `attribution` field crediting them.\n\nNo authentication is required. Only resources their owner has explicitly made public are reachable; everything else returns 404.\n\n**Errors** use one envelope: `{ \"error\": { \"code\": \"...\", \"message\": \"...\" } }`, where `code` is a stable machine-readable string.\n\n**Caching**: every GET returns a weak `ETag`; send it back as `If-None-Match` to get a `304 Not Modified` when the payload is unchanged.\n\n**Pagination**: `/api/v1/cards` is cursor-paginated — follow `next_cursor` until `has_more` is false.\n\n**Rate limiting**: requests are limited per IP (60 requests/minute). Over the limit returns `429` with a `Retry-After` header and error code `rate_limited`.",
    "contact": {
      "name": "Binder"
    }
  },
  "servers": [
...
  "paths": {
...
    "/api/v1/cards": {
      "get": {
        "tags": ["Cards"],
        "summary": "Search the card catalog",
        "description": "Searches Binder's normalized card catalog — every distinct card tracked in Binder, across all three games. Served from Binder's own data, not a live upstream search, so a card appears here once someone tracks it.",
        "operationId": "searchCards",
        "parameters": [
          {
            "name": "game",
            "in": "query",
            "required": false,
            "description": "Restrict results to one game.",
            "schema": { "$ref": "#/components/schemas/Game" }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Case-insensitive substring match on the card name.",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of cards to return in this page (1–100).",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque pagination cursor. Pass the `next_cursor` from the previous page to fetch the next one; omit for the first page.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of matching catalog cards, ordered by name. Follow `next_cursor` while `has_more` is true.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CardsResponse" }
              }
            }
          },
          "400": {
            "description": "Unknown game or invalid cursor.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (60 requests/minute per IP).",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": { "type": "integer" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Game": {
        "type": "string",
        "enum": ["mtg", "pokemon", "onepiece"],
        "description": "The trading card game a binder or card belongs to."
      },
      "Price": {
        "type": "object",
        "description": "Current prices. Any field may be null when unavailable upstream.",
        "properties": {
          "usd": { "type": "number", "nullable": true },
          "usd_foil": { "type": "number", "nullable": true },
          "eur": { "type": "number", "nullable": true }
        }
      },
...
      "CatalogCard": {
        "type": "object",
        "description": "A card in Binder's normalized cross-game model, without holding context (no quantities — the catalog never exposes who holds a card or how many).",
        "properties": {
          "card_id": { "type": "string" },
          "name": { "type": "string" },
          "game": { "$ref": "#/components/schemas/Game" },
          "set_name": { "type": "string" },
          "set_code": { "type": "string" },
          "collector_number": { "type": "string" },
          "rarity": { "type": "string", "nullable": true },
          "image_url": { "type": "string", "format": "uri", "nullable": true, "description": "Absolute URL to the card image (or null). Safe to fetch from any origin." },
          "price": { "$ref": "#/components/schemas/Price" }
        },
        "required": ["card_id", "name", "game", "set_name", "price"]
      },
      "CardsResponse": {
        "type": "object",
        "properties": {
          "count": { "type": "integer", "description": "Number of cards in THIS page (not the catalog total)." },
          "has_more": { "type": "boolean", "description": "Whether another page exists." },
          "next_cursor": {
            "type": "string",
            "nullable": true,
            "description": "Cursor for the next page, or null when has_more is false."
          },
          "cards": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CatalogCard" }
          },
          "attribution": { "type": "string" }
        },
        "required": ["count", "has_more", "next_cursor", "cards", "attribution"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["invalid_request", "not_found", "rate_limited", "internal"],
                "description": "Stable machine-readable error code."
              },
              "message": { "type": "string", "description": "Human-readable detail; may change." }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      }
    }
  }
}